requirements.txt
flask>=3.0.2
mysql-connector-python>=8.3.0
markdown2>=2.4.13